Componente dividido
