+++ frontend/app/(drawer)/explore/types.ts
@@ -1,0 +1,30 @@
+export interface BaseUser {
+  id: number;
+  name: string;
+  username: string;
+  password: string;
+  email: string;
+  phoneNumber?: string;
+  imageProfile?: string;
+}
+
+export interface Artist {
+  id: number;
+  name: string;
+  username: string;
+  email: string;
+  phoneNumber?: string;
+  imageProfile?: string;
+  tableCommisionsPrice?: string;
+  numSlotsOfWork: number;
+  baseUser?: BaseUser;
+}
+
+export interface Work {
+  id: number;
+  name: string;
+  description: string;
+  price: number;
+  artist: Artist;
+  image: string;
+}
